outline: Move the exported outline code into its own file
Keep the unexprted ginkgoNode code in a separate file.
